refactor bot photo commands
